testing/20151116.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20151116</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<h1 id="w3c/testing.20151116.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2015-11-16 / <a href="#end">end</a></h1>
<p>Options: 
<p><a href=20151115>Previous day</a>, <a href=20151117>Next day</a></p>
<ol id=lines>
<li id=l-1 class=hide><a href=#l-1>#</a> Session Start: Mon Nov 16 00:00:00 2015</li>
<li id=l-2 class=hide><a href=#l-2>#</a> Session Ident: #testing</li>
<li id=l-3 class=hide><a href=#l-3>#</a> [00:15] * Quits: ArtB (~ArtB@public.cloak) (&quot;Leaving.&quot;)</li>
<li id=l-4 class=hide><a href=#l-4>#</a> [00:54] * Joins: ArtB (~ArtB@public.cloak)</li>
<li id=l-5 class=hide><a href=#l-5>#</a> [01:01] * Quits: ArtB (~ArtB@public.cloak) (&quot;Leaving.&quot;)</li>
<li id=l-6 class=hide><a href=#l-6>#</a> [02:49] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-7 class=hide><a href=#l-7>#</a> [02:52] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-8 class=hide><a href=#l-8>#</a> [03:14] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-9 class=hide><a href=#l-9>#</a> [04:01] * Quits: hober (~ted@public.cloak) (Client closed connection)</li>
<li id=l-10 class=hide><a href=#l-10>#</a> [04:01] * Joins: hober (~ted@public.cloak)</li>
<li id=l-11 class=hide><a href=#l-11>#</a> [04:05] * Quits: hober (~ted@public.cloak) (hober)</li>
<li id=l-12 class=hide><a href=#l-12>#</a> [04:05] * Joins: hober (~ted@public.cloak)</li>
<li id=l-13 class=hide><a href=#l-13>#</a> [06:24] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-14 class=hide><a href=#l-14>#</a> [07:28] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-15 class=hide><a href=#l-15>#</a> [08:03] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-16 class=hide><a href=#l-16>#</a> [08:03] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-17 class=hide><a href=#l-17>#</a> [08:53] * Joins: rego (~rego@public.cloak)</li>
<li id=l-18 class=hide><a href=#l-18>#</a> [08:57] * Joins: dom (dom@public.cloak)</li>
<li id=l-19 class=hide><a href=#l-19>#</a> [10:00] * Quits: hober (~ted@public.cloak) (Client closed connection)</li>
<li id=l-20 class=hide><a href=#l-20>#</a> [10:01] * Joins: hober (~ted@public.cloak)</li>
<li id=l-21 class=hide><a href=#l-21>#</a> [10:02] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-22 class=hide><a href=#l-22>#</a> [10:04] * Quits: hober (~ted@public.cloak) (Client closed connection)</li>
<li id=l-23 class=hide><a href=#l-23>#</a> [10:05] * Joins: hober (~ted@public.cloak)</li>
<li id=l-24 class=hide><a href=#l-24>#</a> [10:20] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
<li id=l-25 class=hide><a href=#l-25>#</a> [10:21] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-26 class=hide><a href=#l-26>#</a> [10:22] * Joins: dom (dom@public.cloak)</li>
<li id=l-27 class=hide><a href=#l-27>#</a> [11:01] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-28><a href=#l-28>#</a> [11:01] -gitbot:#testing- [wpt-tools] jgraham created manifest_ignore (+1 new commit): <a href="https://github.com/w3c/wpt-tools/commit/0e0863285a92">https://github.com/w3c/wpt-tools/commit/0e0863285a92</a></li>
<li id=l-29><a href=#l-29>#</a> [11:01] -gitbot:#testing- wpt-tools/manifest_ignore 0e08632 James Graham: Ignore testrunner MANIFEST.json file</li>
<li id=l-30 class=hide><a href=#l-30>#</a> [11:01] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-31 class=hide><a href=#l-31>#</a> [11:01] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-32><a href=#l-32>#</a> [11:01] -gitbot:#testing- [wpt-tools] jgraham opened pull request #38: Ignore testrunner MANIFEST.json file (master...manifest_ignore) <a href="https://github.com/w3c/wpt-tools/pull/38">https://github.com/w3c/wpt-tools/pull/38</a></li>
<li id=l-33 class=hide><a href=#l-33>#</a> [11:01] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-34 class=hide><a href=#l-34>#</a> [11:01] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-35><a href=#l-35>#</a> [11:01] -gitbot:#testing- [web-platform-tests] jgraham created tools_ignore (+2 new commits): <a href="https://github.com/w3c/web-platform">https://github.com/w3c/web-platform</a>-tests/compare/0bd1ce1d684e^...3aebf840c7e3</li>
<li id=l-36><a href=#l-36>#</a> [11:01] -gitbot:#testing- web-platform-tests/tools_ignore 0bd1ce1 James Graham: Update tools</li>
<li id=l-37><a href=#l-37>#</a> [11:01] -gitbot:#testing- web-platform-tests/tools_ignore 3aebf84 James Graham: Ignore dirty tools submodule in git status</li>
<li id=l-38 class=hide><a href=#l-38>#</a> [11:01] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-39 class=hide><a href=#l-39>#</a> [11:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-40><a href=#l-40>#</a> [11:02] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #2328: Ignore tools submodule status (master...tools_ignore) <a href="https://github.com/w3c/web-platform-tests/pull/2328">https://github.com/w3c/web-platform-tests/pull/2328</a></li>
<li id=l-41 class=hide><a href=#l-41>#</a> [11:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-42 class=hide><a href=#l-42>#</a> [11:05] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-43><a href=#l-43>#</a> [11:05] -gitbot:#testing- [wpt-tools] Ms2ger closed pull request #38: Ignore testrunner MANIFEST.json file (master...manifest_ignore) <a href="https://github.com/w3c/wpt-tools/pull/38">https://github.com/w3c/wpt-tools/pull/38</a></li>
<li id=l-44 class=hide><a href=#l-44>#</a> [11:05] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-45 class=hide><a href=#l-45>#</a> [11:05] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-46><a href=#l-46>#</a> [11:05] -gitbot:#testing- [wpt-tools] Ms2ger deleted manifest_ignore at 0e08632: <a href="https://github.com/w3c/wpt-tools/commit/0e08632">https://github.com/w3c/wpt-tools/commit/0e08632</a></li>
<li id=l-47 class=hide><a href=#l-47>#</a> [11:05] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-48 class=hide><a href=#l-48>#</a> [11:09] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-49><a href=#l-49>#</a> [11:09] -gitbot:#testing- [web-platform-tests] jgraham force-pushed tools_ignore from 3aebf84 to e58c7f0: <a href="https://github.com/w3c/web-platform-tests/commits/tools_ignore">https://github.com/w3c/web-platform-tests/commits/tools_ignore</a></li>
<li id=l-50><a href=#l-50>#</a> [11:09] -gitbot:#testing- web-platform-tests/tools_ignore 5cac50d James Graham: Update tools submodule</li>
<li id=l-51><a href=#l-51>#</a> [11:09] -gitbot:#testing- web-platform-tests/tools_ignore e58c7f0 James Graham: Ignore dirty tools submodule in git status</li>
<li id=l-52 class=hide><a href=#l-52>#</a> [11:09] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-53 class=hide><a href=#l-53>#</a> [11:45] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-54><a href=#l-54>#</a> [11:45] -gitbot:#testing- [web-platform-tests] Ms2ger deleted tools_ignore at e58c7f0: <a href="https://github.com/w3c/web-platform-tests/commit/e58c7f0">https://github.com/w3c/web-platform-tests/commit/e58c7f0</a></li>
<li id=l-55 class=hide><a href=#l-55>#</a> [11:45] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-56 class=hide><a href=#l-56>#</a> [12:46] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-57><a href=#l-57>#</a> [12:46] -gitbot:#testing- [web-platform-tests] annevk created simple-noreferrer-test (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/2738ea9686fd">https://github.com/w3c/web-platform-tests/commit/2738ea9686fd</a></li>
<li id=l-58><a href=#l-58>#</a> [12:46] -gitbot:#testing- web-platform-tests/simple-noreferrer-test 2738ea9 Anne van Kesteren: Simple rel=noreferrer test</li>
<li id=l-59 class=hide><a href=#l-59>#</a> [12:46] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-60 class=hide><a href=#l-60>#</a> [12:46] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-61><a href=#l-61>#</a> [12:46] -gitbot:#testing- [web-platform-tests] annevk opened pull request #2329: Simple rel=noreferrer test (master...simple-noreferrer-test) <a href="https://github.com/w3c/web-platform-tests/pull/2329">https://github.com/w3c/web-platform-tests/pull/2329</a></li>
<li id=l-62 class=hide><a href=#l-62>#</a> [12:46] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-63 class=hide><a href=#l-63>#</a> [13:05] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-64><a href=#l-64>#</a> [13:05] -gitbot:#testing- [csswg-test] Ms2ger opened pull request #945: Use reftest-wait to ensure background-root-101.xht does not finish early. (master...wait-background-root-101) <a href="https://github.com/w3c/csswg-test/pull/945">https://github.com/w3c/csswg-test/pull/945</a></li>
<li id=l-65 class=hide><a href=#l-65>#</a> [13:05] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-66 class=hide><a href=#l-66>#</a> [13:07] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-67><a href=#l-67>#</a> [13:07] -gitbot:#testing- [csswg-test] Ms2ger created 944-typo (+1 new commit): <a href="https://github.com/w3c/csswg-test/commit/27a21ba009c4">https://github.com/w3c/csswg-test/commit/27a21ba009c4</a></li>
<li id=l-68><a href=#l-68>#</a> [13:07] -gitbot:#testing- csswg-test/944-typo 27a21ba Ms2ger: Fix a typo in universal-selector-002-ref.xht.</li>
<li id=l-69 class=hide><a href=#l-69>#</a> [13:07] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-70 class=hide><a href=#l-70>#</a> [13:07] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-71><a href=#l-71>#</a> [13:07] -gitbot:#testing- [csswg-test] Ms2ger opened pull request #946: Fix a typo in universal-selector-002-ref.xht. (master...944-typo) <a href="https://github.com/w3c/csswg-test/pull/946">https://github.com/w3c/csswg-test/pull/946</a></li>
<li id=l-72 class=hide><a href=#l-72>#</a> [13:07] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-73 class=hide><a href=#l-73>#</a> [13:37] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-74 class=hide><a href=#l-74>#</a> [13:45] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-75><a href=#l-75>#</a> [13:45] -gitbot:#testing- [wpt-tools] jgraham created manifest_seperator (+1 new commit): <a href="https://github.com/w3c/wpt-tools/commit/b1c09a3291d7">https://github.com/w3c/wpt-tools/commit/b1c09a3291d7</a></li>
<li id=l-76><a href=#l-76>#</a> [13:45] -gitbot:#testing- wpt-tools/manifest_seperator b1c09a3 James Graham: Ensure that MANIFEST.json always uses / as a path seperator....</li>
<li id=l-77 class=hide><a href=#l-77>#</a> [13:45] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-78 class=hide><a href=#l-78>#</a> [13:45] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-79><a href=#l-79>#</a> [13:45] -gitbot:#testing- [wpt-tools] jgraham opened pull request #39: Ensure that MANIFEST.json always uses / as a path seperator. (master...manifest_seperator) <a href="https://github.com/w3c/wpt-tools/pull/39">https://github.com/w3c/wpt-tools/pull/39</a></li>
<li id=l-80 class=hide><a href=#l-80>#</a> [13:45] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-81 class=hide><a href=#l-81>#</a> [13:59] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-82 class=hide><a href=#l-82>#</a> [14:16] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-83 class=hide><a href=#l-83>#</a> [15:17] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-84><a href=#l-84>#</a> [15:17] -gitbot:#testing- [wpt-tools] Ms2ger closed pull request #39: Ensure that MANIFEST.json always uses / as a path seperator. (master...manifest_seperator) <a href="https://github.com/w3c/wpt-tools/pull/39">https://github.com/w3c/wpt-tools/pull/39</a></li>
<li id=l-85 class=hide><a href=#l-85>#</a> [15:17] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-86 class=hide><a href=#l-86>#</a> [15:18] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-87><a href=#l-87>#</a> [15:18] -gitbot:#testing- [wpt-tools] Ms2ger deleted manifest_seperator at b1c09a3: <a href="https://github.com/w3c/wpt-tools/commit/b1c09a3">https://github.com/w3c/wpt-tools/commit/b1c09a3</a></li>
<li id=l-88 class=hide><a href=#l-88>#</a> [15:18] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-89 _a="mkwst"><a href=#l-89>#</a> [15:23] &lt;mkwst&gt; jgraham (or whoever): Is there a reasonable shortcut way to test iterations? Like, `assert_array_equals`, but for iterators? :)</li>
<li id=l-90 _a="jgraham" _r="mkwst"><a href=#l-90>#</a> [15:24] &lt;jgraham&gt; mkwst: Iterators weren't even invented when testharness.js was written! Or rather they hadn't made it to js.</li>
<li id=l-91 _a="mkwst" _r="jgraham"><a href=#l-91>#</a> [15:24] &lt;mkwst&gt; jgraham: But they're invented now! And I need to test them!</li>
<li id=l-92 _a="jgraham" _r="mkwst"><a href=#l-92>#</a> [15:24] &lt;jgraham&gt; mkwst: But if you want an assert_every or something, please go ahead</li>
<li id=l-93 _a="jgraham"><a href=#l-93>#</a> [15:25] &lt;jgraham&gt; (idk exactly what the API you have in mind is. Comparing two iterators? compaing an iterator to an array?)</li>
<li id=l-94 _a="mkwst"><a href=#l-94>#</a> [15:25] &lt;mkwst&gt; Comparing an iterator to an array. Same model as `assert_array_equals`. Like, given this iterator, it should produce this output in this order.</li>
<li id=l-95 _a="mkwst"><a href=#l-95>#</a> [15:26] &lt;mkwst&gt; I imagine passing in an array would be a reasonable way of doing that.</li>
<li id=l-96 _a="jgraham" _r="Yep"><a href=#l-96>#</a> [15:26] &lt;jgraham&gt; Yep, sounds entirely reasonable</li>
<li id=l-97 _a="jgraham"><a href=#l-97>#</a> [15:27] &lt;jgraham&gt; If you make a patch, I'll review</li>
<li id=l-98 _a="mkwst"><a href=#l-98>#</a> [15:27] &lt;mkwst&gt; maybe `assert_iteration_equals(thingie, [ '1', '2', '3']))`</li>
<li id=l-99 _a="mkwst"><a href=#l-99>#</a> [15:27] &lt;mkwst&gt; Bleh.</li>
<li id=l-100 _a="mkwst"><a href=#l-100>#</a> [15:27] &lt;mkwst&gt; Ok. Aren't you supposed to write these sorts of things for me? :)</li>
<li id=l-101 _a="jgraham"><a href=#l-101>#</a> [15:28] &lt;jgraham&gt; Well yes, but then you also have to find a reviewer :)</li>
<li id=l-102 _a="mkwst"><a href=#l-102>#</a> [15:29] &lt;mkwst&gt; I'll rubberstamp it! No worries.</li>
<li id=l-103 _a="Ms2ger" _r="mkwst"><a href=#l-103>#</a> [15:29] &lt;Ms2ger&gt; mkwst, any reason not to do `assert_array_equals([...x], y)`?</li>
<li id=l-104 _a="mkwst" _r="Ms2ger"><a href=#l-104>#</a> [15:30] &lt;mkwst&gt; Ms2ger: Does it work?</li>
<li id=l-105 _a="Ms2ger"><a href=#l-105>#</a> [15:30] &lt;Ms2ger&gt; Why wouldn't it?</li>
<li id=l-106 _a="mkwst"><a href=#l-106>#</a> [15:30] &lt;mkwst&gt; `assert_array_equals` compares lengths.</li>
<li id=l-107 _a="mkwst"><a href=#l-107>#</a> [15:30] &lt;mkwst&gt; Which iterators don't have.</li>
<li id=l-108 _a="jgraham" _r="No"><a href=#l-108>#</a> [15:31] &lt;jgraham&gt; No, but [...x] does</li>
<li id=l-109 _a="jgraham"><a href=#l-109>#</a> [15:31] &lt;jgraham&gt; If you have an unbounded iterator, comparing it to an array is problematic in general</li>
<li id=l-110 _a="jgraham"><a href=#l-110>#</a> [15:31] &lt;jgraham&gt; Unless the sematics you want are &quot;compare first N elements&quot;, which is possible but confusing</li>
<li id=l-111 _a="mkwst" _r="Oh"><a href=#l-111>#</a> [15:32] &lt;mkwst&gt; Oh, sorry. I thought you meant &quot;put the array in the first item&quot;.</li>
<li id=l-112 _a="mkwst"><a href=#l-112>#</a> [15:32] &lt;mkwst&gt; Hrm. Maybe that works. I'll check. :)</li>
<li id=l-113 _a="Ms2ger"><a href=#l-113>#</a> [15:32] &lt;Ms2ger&gt; Oh no, I did mean the syntax that's impossible to ever search for :)</li>
<li id=l-114 _a="mkwst" _r="So"><a href=#l-114>#</a> [15:33] &lt;mkwst&gt; So, I think that works for single value iterations, but it doesn't seem to work for multi-value iterations.</li>
<li id=l-115 _a="mkwst" _r="&quot;assert_array_equals"><a href=#l-115>#</a> [15:34] &lt;mkwst&gt; &quot;assert_array_equals: property 0, expected [&quot;a&quot;, &quot;b&quot;] but got [&quot;a&quot;, &quot;b&quot;]&quot; isn't terribly helpful, but I guess it means that the subarrays in the array weren't equal.</li>
<li id=l-116 _a="Ms2ger" _r="Yeah"><a href=#l-116>#</a> [15:36] &lt;Ms2ger&gt; Yeah, you can't use custom equality with it</li>
<li id=l-117 _a="mkwst"><a href=#l-117>#</a> [15:36] &lt;mkwst&gt; assert_object_equals seems to do what I want in this particular case, however.</li>
<li id=l-118 _a="mkwst"><a href=#l-118>#</a> [15:36] &lt;mkwst&gt; I wonder how generalizable it is...</li>
<li id=l-119 _a="Ms2ger"><a href=#l-119>#</a> [15:38] &lt;Ms2ger&gt; I think we're trying to kill assert_object_equals</li>
<li id=l-120 _a="Ms2ger"><a href=#l-120>#</a> [15:38] &lt;Ms2ger&gt; <a href="https://github.com/w3c/web-platform-tests/issues/2033">https://github.com/w3c/web-platform-tests/issues/2033</a></li>
<li id=l-121 _a="mkwst"><a href=#l-121>#</a> [15:39] &lt;mkwst&gt; Ah.</li>
<li id=l-122 _a="mkwst" _r="So"><a href=#l-122>#</a> [15:40] &lt;mkwst&gt; So, if I add a test using it, then MikeSmith will rewrite it for me into something sane? :)</li>
<li id=l-123 _a="jgraham"><a href=#l-123>#</a> [15:44] &lt;jgraham&gt; That is not the right lesson to learn</li>
<li id=l-124 class=hide><a href=#l-124>#</a> [15:45] * Joins: simonstewart (~textual@public.cloak)</li>
<li id=l-125 _a="jgraham"><a href=#l-125>#</a> [15:46] &lt;jgraham&gt; Plus MikeSmith may not have slept for a month and could accidentially rewrite your tests into a form the summons kraken from the deep</li>
<li id=l-126 class=hide><a href=#l-126>#</a> [15:57] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-127 class=hide><a href=#l-127>#</a> [15:58] * Joins: ShaneM (~ShaneM@public.cloak)</li>
<li id=l-128 class=hide><a href=#l-128>#</a> [16:09] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-129><a href=#l-129>#</a> [16:09] -gitbot:#testing- [csswg-test] Ms2ger closed pull request #946: Fix a typo in universal-selector-002-ref.xht. (master...944-typo) <a href="https://github.com/w3c/csswg-test/pull/946">https://github.com/w3c/csswg-test/pull/946</a></li>
<li id=l-130 class=hide><a href=#l-130>#</a> [16:09] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-131 class=hide><a href=#l-131>#</a> [16:10] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-132><a href=#l-132>#</a> [16:10] -gitbot:#testing- [csswg-test] Ms2ger deleted 944-typo at 27a21ba: <a href="https://github.com/w3c/csswg-test/commit/27a21ba">https://github.com/w3c/csswg-test/commit/27a21ba</a></li>
<li id=l-133 class=hide><a href=#l-133>#</a> [16:10] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-134 class=hide><a href=#l-134>#</a> [17:06] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-135 class=hide><a href=#l-135>#</a> [17:21] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-136><a href=#l-136>#</a> [17:21] -gitbot:#testing- [csswg-test] gsnedders closed pull request #945: Use reftest-wait to ensure background-root-101.xht does not finish early. (master...wait-background-root-101) <a href="https://github.com/w3c/csswg-test/pull/945">https://github.com/w3c/csswg-test/pull/945</a></li>
<li id=l-137 class=hide><a href=#l-137>#</a> [17:21] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-138 class=hide><a href=#l-138>#</a> [17:22] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-139><a href=#l-139>#</a> [17:22] -gitbot:#testing- [wptrunner] jgraham pushed 3 new commits to master: <a href="https://github.com/w3c/wptrunner/compare/4510c840152e...24eddf3f49ae">https://github.com/w3c/wptrunner/compare/4510c840152e...24eddf3f49ae</a></li>
<li id=l-140><a href=#l-140>#</a> [17:22] -gitbot:#testing- wptrunner/master 742aa53 Dave Hunt: Use B2G specific ADBDevice constructor, r=bc</li>
<li id=l-141><a href=#l-141>#</a> [17:22] -gitbot:#testing- wptrunner/master ba17768 Ehsan Akhgari: Fix the help string for ./mach web-platform-tests --e10s</li>
<li id=l-142><a href=#l-142>#</a> [17:22] -gitbot:#testing- wptrunner/master 24eddf3 James Graham: Try increasing the reftest timeout in gecko to see if it helps with randomness.</li>
<li id=l-143 class=hide><a href=#l-143>#</a> [17:22] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-144 class=hide><a href=#l-144>#</a> [17:24] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-145><a href=#l-145>#</a> [17:24] -gitbot:#testing- [web-platform-tests] jgraham pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/9718b80985a2350910b4522cb813364dd1081406">https://github.com/w3c/web-platform-tests/commit/9718b80985a2350910b4522cb813364dd1081406</a></li>
<li id=l-146><a href=#l-146>#</a> [17:24] -gitbot:#testing- web-platform-tests/master 9718b80 James Graham: Update to latest wpt-tools</li>
<li id=l-147 class=hide><a href=#l-147>#</a> [17:24] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-148 class=hide><a href=#l-148>#</a> [17:26] * Quits: ShaneM (~ShaneM@public.cloak) (Client closed connection)</li>
<li id=l-149 class=hide><a href=#l-149>#</a> [17:32] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-150 class=hide><a href=#l-150>#</a> [17:39] * Quits: bhill2 (~bhill2@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-151 class=hide><a href=#l-151>#</a> [17:40] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-152 class=hide><a href=#l-152>#</a> [17:56] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-153 class=hide><a href=#l-153>#</a> [17:59] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-154 class=hide><a href=#l-154>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-155><a href=#l-155>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham created sync_2cb71d315f9108b117dc8e73e38919d7a7791d8a (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/2cb71d315f91">https://github.com/w3c/web-platform-tests/commit/2cb71d315f91</a></li>
<li id=l-156><a href=#l-156>#</a> [18:02] -gitbot:#testing- web-platform-tests/sync_2cb71d315f9108b117dc8e73e38919d7a7791d8a 2cb71d3 Boris Zbarsky: Make HTMLCollection check for the element being HTML before checking for its name inside its named getter....</li>
<li id=l-157 class=hide><a href=#l-157>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-158 class=hide><a href=#l-158>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-159><a href=#l-159>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #2330: Make HTMLCollection check for the element being HTML before checking for its name inside its named getter. (master...sync_2cb71d315f9108b117dc8e73e38919d7a7791d8a) <a href="https://github.com/w3c/web-platform-tests/pull/2330">https://github.com/w3c/web-platform-tests/pull/2330</a></li>
<li id=l-160 class=hide><a href=#l-160>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-161 class=hide><a href=#l-161>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-162><a href=#l-162>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham deleted sync_2cb71d315f9108b117dc8e73e38919d7a7791d8a at 2cb71d3: <a href="https://github.com/w3c/web-platform-tests/commit/2cb71d3">https://github.com/w3c/web-platform-tests/commit/2cb71d3</a></li>
<li id=l-163 class=hide><a href=#l-163>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-164 class=hide><a href=#l-164>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-165><a href=#l-165>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham created sync_d21c5d7ad0381ad9aa81d63ce0f6e403d5a1746c (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/d21c5d7ad038">https://github.com/w3c/web-platform-tests/commit/d21c5d7ad038</a></li>
<li id=l-166><a href=#l-166>#</a> [18:02] -gitbot:#testing- web-platform-tests/sync_d21c5d7ad0381ad9aa81d63ce0f6e403d5a1746c d21c5d7 James Graham: Make tests for [before]unload events in window.close not depend on timers,...</li>
<li id=l-167 class=hide><a href=#l-167>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-168 class=hide><a href=#l-168>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-169><a href=#l-169>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #2331: Make tests for [before]unload events in window.close not depend on timers, (master...sync_d21c5d7ad0381ad9aa81d63ce0f6e403d5a1746c) <a href="https://github.com/w3c/web-platform-tests/pull/2331">https://github.com/w3c/web-platform-tests/pull/2331</a></li>
<li id=l-170 class=hide><a href=#l-170>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-171 class=hide><a href=#l-171>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-172><a href=#l-172>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #2331: Make tests for [before]unload events in window.close not depend on timers, (master...sync_d21c5d7ad0381ad9aa81d63ce0f6e403d5a1746c) <a href="https://github.com/w3c/web-platform-tests/pull/2331">https://github.com/w3c/web-platform-tests/pull/2331</a></li>
<li id=l-173 class=hide><a href=#l-173>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-174 class=hide><a href=#l-174>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-175><a href=#l-175>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham deleted sync_d21c5d7ad0381ad9aa81d63ce0f6e403d5a1746c at d21c5d7: <a href="https://github.com/w3c/web-platform-tests/commit/d21c5d7">https://github.com/w3c/web-platform-tests/commit/d21c5d7</a></li>
<li id=l-176 class=hide><a href=#l-176>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-177 class=hide><a href=#l-177>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-178><a href=#l-178>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham created sync_cf3ef1ff33b3a3747261811326854ca962360c78 (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/cf3ef1ff33b3">https://github.com/w3c/web-platform-tests/commit/cf3ef1ff33b3</a></li>
<li id=l-179><a href=#l-179>#</a> [18:02] -gitbot:#testing- web-platform-tests/sync_cf3ef1ff33b3a3747261811326854ca962360c78 cf3ef1f James Graham: Wait for at least one close event to fire before ending close-multiple test, to prevent instability,...</li>
<li id=l-180 class=hide><a href=#l-180>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-181 class=hide><a href=#l-181>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-182><a href=#l-182>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #2332: Wait for at least one close event to fire before ending close-multiple test, to prevent instability, (master...sync_cf3ef1ff33b3a3747261811326854ca962360c78) <a href="https://github.com/w3c/web-platform-tests/pull/2332">https://github.com/w3c/web-platform-tests/pull/2332</a></li>
<li id=l-183 class=hide><a href=#l-183>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-184 class=hide><a href=#l-184>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-185><a href=#l-185>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham deleted sync_cf3ef1ff33b3a3747261811326854ca962360c78 at cf3ef1f: <a href="https://github.com/w3c/web-platform-tests/commit/cf3ef1f">https://github.com/w3c/web-platform-tests/commit/cf3ef1f</a></li>
<li id=l-186 class=hide><a href=#l-186>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-187 class=hide><a href=#l-187>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-188><a href=#l-188>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham created sync_df11d647cafc87edc0e821f1d35b7e790fc26046 (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/df11d647cafc">https://github.com/w3c/web-platform-tests/commit/df11d647cafc</a></li>
<li id=l-189><a href=#l-189>#</a> [18:02] -gitbot:#testing- web-platform-tests/sync_df11d647cafc87edc0e821f1d35b7e790fc26046 df11d64 James Graham: fixup! Wait for at least one close event to fire before ending close-multiple test, to prevent instability, rs=Ms2ger...</li>
<li id=l-190 class=hide><a href=#l-190>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-191 class=hide><a href=#l-191>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-192><a href=#l-192>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #2333: fixup! Wait for at least one close event to fire before ending close-multiple test, to prevent instability, rs=Ms2ger (master...sync_df11d647cafc87edc0e821f1d35b7e790fc26046) <a href="https://github.com/w3c/web-platform-tests/pull/2333">https://github.com/w3c/web-platform-tests/pull/2333</a></li>
<li id=l-193 class=hide><a href=#l-193>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-194 class=hide><a href=#l-194>#</a> [18:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-195><a href=#l-195>#</a> [18:02] -gitbot:#testing- [web-platform-tests] jgraham deleted sync_df11d647cafc87edc0e821f1d35b7e790fc26046 at df11d64: <a href="https://github.com/w3c/web-platform-tests/commit/df11d64">https://github.com/w3c/web-platform-tests/commit/df11d64</a></li>
<li id=l-196 class=hide><a href=#l-196>#</a> [18:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-197 class=hide><a href=#l-197>#</a> [18:03] * Joins: kegan (~kegan@public.cloak)</li>
<li id=l-198 _a="kegan"><a href=#l-198>#</a> [18:04] &lt;kegan&gt; test</li>
<li id=l-199 _a="Ms2ger"><a href=#l-199>#</a> [18:04] &lt;Ms2ger&gt; pass</li>
<li id=l-200 class=hide><a href=#l-200>#</a> [18:05] * Quits: M-hrjet (~hrjetmatri@public.cloak) (Client closed connection)</li>
<li id=l-201 class=hide><a href=#l-201>#</a> [18:06] * Joins: MatrixBridge (~matrixirc@public.cloak)</li>
<li id=l-202 class=hide><a href=#l-202>#</a> [18:06] * Joins: M-hrjet (~hrjetmatri@public.cloak)</li>
<li id=l-203 class=hide><a href=#l-203>#</a> [18:06] * Quits: MatrixBridge (~matrixirc@public.cloak) (Client closed connection)</li>
<li id=l-204 class=hide><a href=#l-204>#</a> [18:06] * Quits: M-hrjet (~hrjetmatri@public.cloak) (Client closed connection)</li>
<li id=l-205 class=hide><a href=#l-205>#</a> [18:08] * Joins: M-hrjet (~hrjetmatri@public.cloak)</li>
<li id=l-206 _a="kegan"><a href=#l-206>#</a> [18:09] &lt;kegan&gt; random test</li>
<li id=l-207 _a="Ms2ger"><a href=#l-207>#</a> [18:10] &lt;Ms2ger&gt; intermittent fail</li>
<li id=l-208 _a="kegan"><a href=#l-208>#</a> [18:11] &lt;kegan&gt; :)</li>
<li id=l-209 class=hide><a href=#l-209>#</a> [18:35] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-210 class=hide><a href=#l-210>#</a> [18:39] * Quits: simonstewart (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-211 class=hide><a href=#l-211>#</a> [18:44] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-212 class=hide><a href=#l-212>#</a> [18:51] * Quits: scheib (~sid4467@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-213 class=hide><a href=#l-213>#</a> [18:52] * Joins: scheib (~sid4467@public.cloak)</li>
<li id=l-214 class=hide><a href=#l-214>#</a> [18:56] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-215><a href=#l-215>#</a> [18:56] -gitbot:#testing- [web-platform-tests] annevk force-pushed simple-noreferrer-test from 2738ea9 to a5b5408: <a href="https://github.com/w3c/web-platform-tests/commits/simple-noreferrer-test">https://github.com/w3c/web-platform-tests/commits/simple-noreferrer-test</a></li>
<li id=l-216><a href=#l-216>#</a> [18:56] -gitbot:#testing- web-platform-tests/simple-noreferrer-test a5b5408 Anne van Kesteren: Simple rel=noreferrer test</li>
<li id=l-217 class=hide><a href=#l-217>#</a> [18:56] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-218 class=hide><a href=#l-218>#</a> [18:57] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
<li id=l-219 class=hide><a href=#l-219>#</a> [19:10] * Parts: kegan (~kegan@public.cloak) (Leaving)</li>
<li id=l-220 class=hide><a href=#l-220>#</a> [19:11] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-221 class=hide><a href=#l-221>#</a> [19:18] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-222 class=hide><a href=#l-222>#</a> [19:21] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-223><a href=#l-223>#</a> [19:21] -gitbot:#testing- [web-platform-tests] annevk created storage-document-domain (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/ac4fb3819512">https://github.com/w3c/web-platform-tests/commit/ac4fb3819512</a></li>
<li id=l-224><a href=#l-224>#</a> [19:21] -gitbot:#testing- web-platform-tests/storage-document-domain ac4fb38 Anne van Kesteren: localStorage and document.domain test...</li>
<li id=l-225 class=hide><a href=#l-225>#</a> [19:21] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-226 class=hide><a href=#l-226>#</a> [19:21] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-227><a href=#l-227>#</a> [19:21] -gitbot:#testing- [web-platform-tests] annevk opened pull request #2334: localStorage and document.domain test (master...storage-document-domain) <a href="https://github.com/w3c/web-platform-tests/pull/2334">https://github.com/w3c/web-platform-tests/pull/2334</a></li>
<li id=l-228 class=hide><a href=#l-228>#</a> [19:21] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-229 class=hide><a href=#l-229>#</a> [19:31] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;nn&quot;)</li>
<li id=l-230 class=hide><a href=#l-230>#</a> [20:28] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-231 class=hide><a href=#l-231>#</a> [20:32] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-232 class=hide><a href=#l-232>#</a> [20:32] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-233 class=hide><a href=#l-233>#</a> [20:39] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-234 class=hide><a href=#l-234>#</a> [20:41] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-235 class=hide><a href=#l-235>#</a> [20:49] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-236 class=hide><a href=#l-236>#</a> [20:52] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-237 class=hide><a href=#l-237>#</a> [20:54] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-238 class=hide><a href=#l-238>#</a> [21:03] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-239 class=hide><a href=#l-239>#</a> [21:05] * Quits: rniwa (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-240 class=hide><a href=#l-240>#</a> [21:17] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-241 class=hide _a="terri_offline"><a href=#l-241>#</a> [21:23] * terri_offline is now known as terri</li>
<li id=l-242 class=hide><a href=#l-242>#</a> [21:29] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-243 class=hide><a href=#l-243>#</a> [21:31] * Joins: simonstewart (~textual@public.cloak)</li>
<li id=l-244 class=hide><a href=#l-244>#</a> [21:55] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-245 class=hide><a href=#l-245>#</a> [21:55] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-246 class=hide><a href=#l-246>#</a> [21:57] * Quits: simonstewart (~textual@public.cloak) (&quot;Textual IRC Client: www.textualapp.com&quot;)</li>
<li id=l-247 class=hide><a href=#l-247>#</a> [22:00] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-248 class=hide><a href=#l-248>#</a> [22:01] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-249 class=hide><a href=#l-249>#</a> [22:01] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-250 class=hide><a href=#l-250>#</a> [22:03] * Quits: rego (~rego@public.cloak) (Client closed connection)</li>
<li id=l-251 class=hide><a href=#l-251>#</a> [22:03] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-252 class=hide><a href=#l-252>#</a> [22:04] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-253 class=hide><a href=#l-253>#</a> [22:07] * Joins: bhill2_ (~bhill2@public.cloak)</li>
<li id=l-254 class=hide><a href=#l-254>#</a> [22:07] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-255 class=hide><a href=#l-255>#</a> [22:08] * Joins: zcorpan_ (~zcorpan@public.cloak)</li>
<li id=l-256 class=hide><a href=#l-256>#</a> [22:09] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-257 class=hide _a="terri"><a href=#l-257>#</a> [22:17] * terri is now known as terri_offline</li>
<li id=l-258 class=hide><a href=#l-258>#</a> [22:17] * Quits: bhill2_ (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-259 class=hide><a href=#l-259>#</a> [22:29] * Quits: zcorpan_ (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-260 class=hide><a href=#l-260>#</a> [22:30] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-261 class=hide _a="terri_offline"><a href=#l-261>#</a> [22:31] * terri_offline is now known as terri</li>
<li id=l-262 class=hide><a href=#l-262>#</a> [22:56] * Quits: rniwa (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-263 class=hide><a href=#l-263>#</a> [23:15] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-264 class=hide _a="terri"><a href=#l-264>#</a> [23:30] * terri is now known as terri_offline</li>
<li id=l-265 class=hide><a href=#l-265>#</a> Session Close: Tue Nov 17 00:00:00 2015</li>
</ol>
<p id=end><a href=20151115>Previous day</a>, <a href=20151117>Next day</a></p>
<p>Think <a href="/irc-logs/">these logs</a> are useful? Then <a href="https://www.paypal.com/cgi-bin/webscr?cmd=_donations&amp;business=PMXKLAEH53JSS&amp;lc=GB&amp;item_name=Krijn%20Hoetmer&amp;item_number=IRC%20Logs&amp;currency_code=EUR&amp;bn=PP%2dDonationsBF%3abtn_donate_SM%2egif%3aNonHosted"><strong>please donate</strong></a> to show your gratitude (and keep them up, of course). <a href="https://twitter.com/krijnhoetmer/status/414824591337734144">Thanks! — Krijn</a></p>
<script src=/irc-logs/s2.js></script>
